Track book stats and update them dynamically
In order to keep the library more helpful for the user, the stats bar
was implemented with the number of all/read/unread books.

If the user potentially adds more than 100 books, then it is going to
be difficult for them to keep track of how many books there are in the
library. So instead of them counting the numbers manually, the program
will give them the information and update it in real time. As a result,
UX will be singificantly improved.